Adding update that solves one logger issue
